backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.0
SQLAlchemy==2.0.0
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0
twilio==8.10.0
APScheduler==3.10.0
